% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{RchatSF}
\alias{RchatSF}
\title{DeepSeek Chat Interface}
\usage{
RchatSF(api_key, model = "deepseek-ai/DeepSeek-R1")
}
\arguments{
\item{api_key}{DeepSeek API key}

\item{model}{model name (e.g. "deepseek-ai/DeepSeek-R1")}
}
\description{
DeepSeek Chat Interface
}
